temp: Enable Celery distributed tracing in stage edxapp
Datadog Support suggested that we try enabling this to see if it changes
anything about the orphaned spans on the edxapp workers (unexpected
top-level spans that are not `operation_name:celery.run`, but instead are
other spans that have lost their parent association).

Just going to enable this on stage so we don't mess with traces in prod,
for now.

See https://github.com/edx/edx-arch-experiments/issues/822