feat: add support for installing extra npm packages into edxapp
This commit adds support for installing extra npm packages in edxapp to the edxapp
deploy playbook. In order to add an extra npm package to edxapp, add the package to
the EDXAPP_PRIVATE_NPM_REQUIREMENTS list.

The format of the package should be as follows. name is required and version is optional.
version can be an explicit version or a version range, as supported by npm.

EDXAPP_PRIVATE_NPM_REQUIREMENTS:
  - name: mypackage
    version: 1.0.1
  - name: git+https://git.myproject.org/MyProject#egg=MyProject