fix: Remove call to print WEBPACK_CONFIG_PATH when updating static assets
This setting was removed from edxapp base configs in
https://github.com/openedx/edx-platform/pull/36494 and one of these lines
started failing with "CommandError: WEBPACK_CONFIG_PATH not found in
settings."

We don't actually need to print any *particular* setting here; see comment
for why we're making this call at all. So here I change it to use a
setting that is more unlikely to go away any time soon.

Also, add some "comments" to help draw attention to the comment higher up,
as well as a little more explanation of the situation.